requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Mail>=0.9
Flask-WTF>=1.2
Flask-Login>=0.6
python-dotenv>=1.0
python-dateutil>=2.8
PyMySQL>=1.1
SQLAlchemy>=2.0
alembic>=1.13

pytest>=7.4
hypothesis>=6.92
